clean app and improve project tree
